+++ src/Bootloader/Auth/TokenStorage/SessionTokensBootloader.php
@@ -1,0 +1,33 @@
+<?php
+
+/**
+ * Spiral Framework.
+ *
+ * @license   MIT
+ * @author    Anton Titov (Wolfy-J)
+ */
+
+declare(strict_types=1);
+
+namespace Spiral\Bootloader\Auth\TokenStorage;
+
+use Spiral\Auth\Session\TokenStorage as SessionStorage;
+use Spiral\Auth\TokenStorageInterface;
+use Spiral\Boot\Bootloader\Bootloader;
+use Spiral\Bootloader\Auth\HttpAuthBootloader;
+use Spiral\Bootloader\Http\SessionBootloader;
+
+/**
+ * Stores authentication token in session.
+ */
+final class SessionTokensBootloader extends Bootloader
+{
+    protected const DEPENDENCIES = [
+        HttpAuthBootloader::class,
+        SessionBootloader::class
+    ];
+
+    protected const SINGLETONS = [
+        TokenStorageInterface::class => SessionStorage::class
+    ];
+}