+++ tests/app/config/database.php
@@ -1,13 +1,14 @@
 <?php
 
-declare(strict_types=1);
-
 /**
  * Spiral Framework.
  *
  * @license   MIT
  * @author    Anton Titov (Wolfy-J)
  */
+
+declare(strict_types=1);
+
 use Spiral\Database\Driver;
 
 return [